Refactor dataloading utilities and cleanup imports
